+++ Assets/Scripts/Gameplay/Models/Configurations/CameraConfig.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3034c0f701ddc234f885df5adf1cc64c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
